+++ src/index.react-native.ts
@@ -1,0 +1,2 @@
+export * from './create';
+export * from './persist/index.rn';