+++ src/persist/index.rn.ts
@@ -1,8 +1,7 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import { getPersistStore, debounce } from './get-persist';
 
-export const isClient = true;
-
-export default {
+const cacheStore = {
   setItem(key: string, state: unknown) {
     return AsyncStorage.setItem(key, JSON.stringify(state));
   },
@@ -12,7 +11,7 @@ export default {
       try {
         return JSON.parse(result);
       } catch (e) {
-        AsyncStorage.removeItem(key).catch((e) => {
+        AsyncStorage.removeItem(key).catch(() => {
           //
         });
       }
@@ -23,3 +22,8 @@ export default {
     return AsyncStorage.removeItem(key);
   },
 };
+
+export const isClient = true;
+const { wrapState, persistStore } = getPersistStore(cacheStore, isClient);
+
+export { debounce, wrapState, persistStore };